refactor(frontend): split room message stores
